Zoom in and out with (shift key) double-click
