+++ demo-timeline/src/timeline-chart.ts
@@ -1,5 +1,6 @@
 import Chart from 'chart.js';
 import 'chartjs-plugin-zoom';
+import { applyChartInstanceOverrides } from './chartjs-horizontal-bar';
 import { Level, LevelParsed } from '../../src/types/level';
 import { MediaPlaylist } from '../../src/types/media-playlist';
 import { TrackSet } from '../../src/types/track';
@@ -31,8 +32,10 @@ export class TimelineChart {
       }]
     });
 
+    applyChartInstanceOverrides(chart);
+
     // Log object on click and seek to position
-    canvas.onclick = (event) => {
+    canvas.onclick = (event: MouseEvent) => {
       const element = chart.getElementAtEvent(event);
       if (element.length) {
         const dataset = chart.data.datasets[element[0]._datasetIndex];
@@ -47,22 +50,35 @@ export class TimelineChart {
       }
     };
 
-    canvas.ondblclick = function () {
-      chart.resetZoom();
+    canvas.ondblclick = (event: MouseEvent) => {
+      const chartArea: { left, top, right, bottom } = chart.chartArea;
+      const element = chart.getElementAtEvent(event);
+      const pos = Chart.helpers.getRelativePosition(event, chart);
+      const scale = chart.scales['x-axis-0'];
+      const range = scale.max - scale.min;
+      const newDiff = range * (event.getModifierState('Shift') ? -1.0 : 0.5);
+      const minPercent = (scale.getValueForPixel(pos.x) - scale.min) / range;
+      const maxPercent = 1 - minPercent;
+      const minDelta = newDiff * minPercent;
+      const maxDelta = newDiff * maxPercent;
+      // zoom in when double clicking near elements in chart area
+      if (element.length || pos.x > chartArea.left) {
+        scale.options.ticks.min = Math.max(this.minZoom, scale.min + minDelta);
+        scale.options.ticks.max = Math.min(this.maxZoom, scale.max - maxDelta);
+      } else {
+        // chart.resetZoom();
+        scale.options.ticks.min = this.minZoom;
+        scale.options.ticks.max = this.maxZoom;
+      }
+      this.update();
     };
 
     // TODO: Prevent zoom over y axis labels
-
-    // Parse custom dataset (Fragment)
-    Object.keys(chart.scales).forEach((axis) => {
-      const scale = chart.scales[axis];
-      scale._parseValue = scaleParseValue;
-    });
   }
 
   reset () {
     const { labels, datasets } = this.chart.data;
-
+    this.maxZoom = 60;
     labels.length = 0;
     datasets.length = 0;
     this.resize(datasets);
@@ -157,13 +173,33 @@ export class TimelineChart {
     const { data } = datasets.find(dataset => dataset.url === url);
     data.length = 0;
     details.fragments.forEach((fragment) => {
+      // TODO: keep track of initial playlist start and duration so that we can show drift and pts offset
+      // (Make that a feature of hls.js v1.0.0 fragments)
       data.push(Object.assign({}, fragment));
     });
+    this.maxZoom = Math.max(totalduration, this.maxZoom);
+    this.rafDebounceRequestId = self.requestAnimationFrame(() => this.update());
+  }
+
+  get minZoom (): number {
+    if (this.chart.config?.options?.plugins?.zoom?.zoom?.rangeMin) {
+      return this.chart.config.options.plugins.zoom.zoom.rangeMin.x;
+    }
+    return 60;
+  }
+
+  get maxZoom (): number {
     if (this.chart.config?.options?.plugins?.zoom?.zoom?.rangeMax) {
-      this.chart.config.options.plugins.zoom.zoom.rangeMax.x = Math.max(totalduration,
+      return this.chart.config.options.plugins.zoom.zoom.rangeMax.x;
+    }
+    return 60;
+  }
+
+  set maxZoom (x: number) {
+    if (this.chart.config?.options?.plugins?.zoom?.zoom?.rangeMax) {
+      this.chart.config.options.plugins.zoom.zoom.rangeMax.x = Math.max(x,
         this.chart.config.options.plugins.zoom.zoom.rangeMax.x);
     }
-    this.rafDebounceRequestId = self.requestAnimationFrame(() => this.update());
   }
 
   updateFragment (data: FragChangedData) {
@@ -201,7 +237,12 @@ export class TimelineChart {
         sourceBuffer
       });
       sourceBuffer.onupdate = () => {
-        replaceTimeRangeTuples(sourceBuffer.buffered, data);
+        try {
+          replaceTimeRangeTuples(sourceBuffer.buffered, data);
+        } catch (error) {
+          console.warn(error);
+          return;
+        }
         replaceTimeRangeTuples(media.buffered, mediaBufferData);
         this.update();
       };
@@ -350,158 +391,5 @@ function getChartOptions () {
         }
       }
     }
-  };
-}
-
-// Modify horizontalBar so that each dataset (fragments, timeRanges) draws on the same row (level, track or buffer)
-Chart.controllers.horizontalBar.prototype.calculateBarValuePixels = function (datasetIndex, index, options) {
-  const chart = this.chart;
-  const scale = this._getValueScale();
-  const datasets = chart.data.datasets;
-  // const metasets = scale._getMatchingVisibleMetas(this._type);
-  const value = scale._parseValue(datasets[datasetIndex].data[index]);
-  const start = value.start === undefined ? 0 : value.max >= 0 && value.min >= 0 ? value.min : value.max;
-  const length = value.start === undefined ? value.end : value.max >= 0 && value.min >= 0 ? value.max - value.min : value.min - value.max;
-  const base = scale.getPixelForValue(start);
-  const head = scale.getPixelForValue(start + length);
-  const size = head - base;
-
-  return {
-    size: size,
-    base: base,
-    head: head,
-    center: head + size / 2
-  };
-};
-
-Chart.controllers.horizontalBar.prototype.calculateBarIndexPixels = function (datasetIndex, index, ruler, options) {
-  const rowHeight = 35;
-  const size = rowHeight * options.categoryPercentage;
-  const center = datasetIndex * rowHeight + (rowHeight / 2);
-  return {
-    base: center - size / 2,
-    head: center + size / 2,
-    center,
-    size
-  };
-};
-
-Chart.controllers.horizontalBar.prototype.draw = function () {
-  const chart = this.chart;
-  const scale = this._getValueScale();
-  const rects = this.getMeta().data;
-  const dataset = this.getDataset();
-  const len = rects.length;
-  if (len !== dataset.data.length) {
-    // View does not match dataset (wait for redraw)
-    return;
-  }
-  const ctx: CanvasRenderingContext2D = chart.ctx;
-  const chartArea: { left, top, right, bottom } = chart.chartArea;
-  Chart.helpers.canvas.clipArea(ctx, chartArea);
-  const lineHeight = Math.ceil(ctx.measureText('0').actualBoundingBoxAscent) + 2;
-  for (let i = 0; i < len; ++i) {
-    const rect = rects[i];
-    const view = rect._view;
-    if (!intersects(view.base, view.x, chartArea.left, chartArea.right)) {
-      // Do not draw elements outside of the chart's viewport
-      continue;
-    }
-    const obj = dataset.data[i];
-    scale._parseValue = scaleParseValue;
-    const val = scale._parseValue(obj);
-    if (!isNaN(val.min) && !isNaN(val.max)) {
-      const { stats } = obj;
-      const isFragment = !!stats;
-      const bounds = boundingRects(view);
-      const drawText = bounds.w > lineHeight;
-      if (isFragment) {
-        if (drawText) {
-          view.borderWidth = 1;
-          if (i === 0) {
-            view.borderSkipped = null;
-          }
-        } else {
-          view.borderWidth = 0;
-          view.backgroundColor = `rgba(0, 0, 0, ${0.1 + (i % 2) / 4})`;
-        }
-      }
-      rect.draw();
-      if (isFragment) {
-        if (stats.aborted) {
-          ctx.fillStyle = 'rgba(100, 0, 0, 0.3)';
-          ctx.fillRect(bounds.x, bounds.y, bounds.w, bounds.h);
-        }
-        if (stats.loaded && stats.total) {
-          ctx.fillStyle = 'rgba(50, 20, 100, 0.3)';
-          ctx.fillRect(bounds.x, bounds.y, bounds.w * stats.loaded / stats.total, bounds.h);
-        }
-      }
-      if (drawText) {
-        const start = val.start; // obj.start;
-        ctx.fillStyle = 'rgb(0, 0, 0)';
-        if (stats) {
-          const snLabel = 'sn: ' + obj.sn;
-          const textWidth = Math.min(ctx.measureText(snLabel).width + 2, bounds.w - 2);
-          ctx.fillText(snLabel, bounds.x + bounds.w - textWidth, bounds.y + lineHeight, bounds.w - 4);
-        }
-        const float = start !== (start | 0);
-        const fixedPoint = float ? Math.min(5, Math.max(1, Math.floor(bounds.w / 10 - 1))) : 0;
-        const startString = fixedPoint ? start.toFixed(fixedPoint).replace(/\.0$/, '..') : start.toString();
-        ctx.fillText(startString, bounds.x + 2, bounds.y + bounds.h - 3, bounds.w - 5);
-      }
-    }
-  }
-
-  Chart.helpers.canvas.unclipArea(chart.ctx);
-};
-
-function scaleParseValue (value: number[] | any) {
-  let start, end, min, max;
-
-  if (value === undefined) {
-    console.warn('Chart values undefined (update chart)');
-    return {};
-  }
-
-  if (Array.isArray(value)) {
-    start = +this.getRightValue(value[0]);
-    end = +this.getRightValue(value[1]);
-    min = Math.min(start, end);
-    max = Math.max(start, end);
-  } else {
-    start = +this.getRightValue(value.start);
-    if ('end' in value) {
-      end = +this.getRightValue(value.end);
-    } else {
-      end = +this.getRightValue(value.start + value.duration);
-    }
-    min = Math.min(start, end);
-    max = Math.max(start, end);
-  }
-
-  return {
-    min,
-    max,
-    start,
-    end
-  };
-}
-
-function intersects (x1, x2, x3, x4) {
-  return x2 > x3 && x1 < x4;
-}
-
-function boundingRects (vm) {
-  const half = vm.height / 2;
-  const left = Math.min(vm.x, vm.base);
-  const right = Math.max(vm.x, vm.base);
-  const top = vm.y - half;
-  const bottom = vm.y + half;
-  return {
-    x: left,
-    y: top,
-    w: right - left,
-    h: bottom - top
   };
 }
